Changed design of ul in PayView

+++ click.parent/click.client/www/app/view/View.PayConfirm.tag
@@ -65,8 +65,10 @@
             riot.mount('view-service-pincards');
         }
 
-        history.arrayOfHistory.push('view-pay-confirm');
-        sessionStorage.setItem('history', JSON.stringify(history.arrayOfHistory));
+        if(history.arrayOfHistory[history.arrayOfHistory.length - 1] != 'view-pay-confirm') {
+            history.arrayOfHistory.push('view-pay-confirm');
+            sessionStorage.setItem('history', JSON.stringify(history.arrayOfHistory));
+        }
 
         scope.servicesMap = JSON.parse(localStorage.getItem("click_client_servicesMap"));
         scope.categoryNamesMap = JSON.parse(localStorage.getItem("click_client_categoryNamesMap"));
